- Major refactoring operation to support PRIDE projects - PRIDE support not complete! - Created new Netbeans derived GUI with logger - Enabled PeptideShaker processing via PeptideShakerCLI

+++ compomics-relims/src/main/java/com/compomics/relims/model/beans/RelimsProjectBean.java
@@ -66,7 +66,7 @@ public class RelimsProjectBean {
         return iProjectID;
     }
 
-    public void setProjectID(int aProjectID) {
+    public void setProjectID(long aProjectID) {
         iProjectID = aProjectID;
     }
 }
